feat: organize files layout
